DataHarvester/services/scraper_service/pyproject.toml
# Software/DataHarvester/services/scraper_service/pyproject.toml

[tool.black]
line-length = 135
target-version = ['py312']
extend-exclude = '''
/(
    \.venv
  | \.mypy_cache
  | \.pytest_cache
  | \.git
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 135

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scraper-service"
version = "0.1.0"
description = "Data Ingestion Service for DataHarvester"
requires-python = "==3.12.8"
dependencies = [
    "yt-dlp>=2023.11.16",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "pymongo>=4.6.1",
    "ffmpeg-python>=0.2.0",
    "pytube>=15.0.0",
    "youtube-transcript-api>=0.6.1",
    "feature-engine>=1.6.0",
    "pyjanitor>=0.26.0",
    "nltk>=3.8.1",
    "spacy>=3.7.2",
    "presidio-analyzer>=2.2.33",
    "presidio-anonymizer>=2.2.33",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pytest>=7.4.0"
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-mock>=3.10.0"
]

[tool.setuptools]
packages = [
    "application",
    "application.services",
    "application.services.text",
    "application.services.transcript",
    "domain",
    "domain.exceptions",
    "domain.models",
    "infrastructure",
    "infrastructure.config",
    "infrastructure.error_handling",
    "infrastructure.logging",
    "infrastructure.monitoring",
    "infrastructure.redis",
    "presentation",
    "presentation.cli",
    "validation",
    "validation.rules",
    "validation.validators"
]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests"
]